+++ lib/service/auth2.dart
@@ -214,6 +214,7 @@ class Auth2 with Service, NetworkAuthProvider implements NotificationsListener {
   Auth2LoginType get oidcLoginType => Auth2LoginType.oidcIllinois;
   Auth2LoginType get phoneLoginType => Auth2LoginType.phoneTwilio;
   Auth2LoginType get emailLoginType => Auth2LoginType.email;
+  Auth2LoginType get usernameLoginType => Auth2LoginType.username;
 
   Auth2Token? get token => _token ?? _anonymousToken;
   Auth2Token? get userToken => _token;
@@ -230,14 +231,17 @@ class Auth2 with Service, NetworkAuthProvider implements NotificationsListener {
   bool get isOidcLoggedIn => (_account?.authType?.loginType == oidcLoginType);
   bool get isPhoneLoggedIn => (_account?.authType?.loginType == phoneLoginType);
   bool get isEmailLoggedIn => (_account?.authType?.loginType == emailLoginType);
+  bool get isUsernameLoggedIn => (_account?.authType?.loginType == usernameLoginType);
 
   bool get isOidcLinked => _account?.isAuthTypeLinked(oidcLoginType) ?? false;
   bool get isPhoneLinked => _account?.isAuthTypeLinked(phoneLoginType) ?? false;
   bool get isEmailLinked => _account?.isAuthTypeLinked(emailLoginType) ?? false;
+  bool get isUsernameLinked => _account?.isAuthTypeLinked(usernameLoginType) ?? false;
 
   List<Auth2Type> get linkedOidc => _account?.getLinkedForAuthType(oidcLoginType) ?? [];
   List<Auth2Type> get linkedPhone => _account?.getLinkedForAuthType(phoneLoginType) ?? [];
   List<Auth2Type> get linkedEmail => _account?.getLinkedForAuthType(emailLoginType) ?? [];
+  List<Auth2Type> get linkedUsername => _account?.getLinkedForAuthType(usernameLoginType) ?? [];
 
   bool get hasUin => (0 < (uin?.length ?? 0));
   String? get uin => _account?.authType?.uiucUser?.uin;
@@ -247,6 +251,7 @@ class Auth2 with Service, NetworkAuthProvider implements NotificationsListener {
   String? get firstName => StringUtils.ensureNotEmpty(profile?.firstName, defaultValue: _account?.authType?.uiucUser?.firstName ?? '');
   String? get email => StringUtils.ensureNotEmpty(profile?.email, defaultValue: _account?.authType?.uiucUser?.email ?? '');
   String? get phone => StringUtils.ensureNotEmpty(profile?.phone, defaultValue: _account?.authType?.phone ?? '');
+  String? get username => _account?.username;
 
   bool get isEventEditor => hasRole("event approvers");
   bool get isStadiumPollManager => hasRole("stadium poll manager");
@@ -727,6 +732,113 @@ class Auth2 with Service, NetworkAuthProvider implements NotificationsListener {
       return (response?.statusCode == 200);
     }
     return false;
+  }
+
+  // Username Authentication
+
+  Future<Auth2UsernameSignInResult> authenticateWithUsername(String? username, String? password) async {
+    if ((Config().coreUrl != null) && (Config().appPlatformId != null) && (Config().coreOrgId != null) && (username != null) && (password != null)) {
+
+      NotificationService().notify(notifyLoginStarted, usernameLoginType);
+
+      String url = "${Config().coreUrl}/services/auth/login";
+      Map<String, String> headers = {
+        'Content-Type': 'application/json'
+      };
+      String? post = JsonUtils.encode({
+        'auth_type': auth2LoginTypeToString(usernameLoginType),
+        'app_type_identifier': Config().appPlatformId,
+        'api_key': Config().rokwireApiKey,
+        'org_id': Config().coreOrgId,
+        'creds': {
+          "username": username,
+          "password": password
+        },
+        'params': {
+          "sign_up": false,
+        },
+        'profile': _anonymousProfile?.toJson(),
+        'preferences': _anonymousPrefs?.toJson(),
+        'device': deviceInfo,
+      });
+
+      Response? response = await Network().post(url, headers: headers, body: post);
+      if (response?.statusCode == 200) {
+        bool result = await processLoginResponse(JsonUtils.decodeMap(response?.body));
+        _notifyLogin(usernameLoginType, result);
+        return result ? Auth2UsernameSignInResult.succeeded : Auth2UsernameSignInResult.failed;
+      }
+      else {
+        _notifyLogin(usernameLoginType, false);
+        Auth2Error? error = Auth2Error.fromJson(JsonUtils.decodeMap(response?.body));
+        if (error?.status == 'not-found') {
+          return Auth2UsernameSignInResult.failedNotFound;
+        } else if (error?.status == 'invalid') {
+          return Auth2UsernameSignInResult.failedInvalid;
+        }
+      }
+    }
+    return Auth2UsernameSignInResult.failed;
+  }
+
+  Future<Auth2UsernameSignUpResult> signUpWithUsername(String? username, String? password) async {
+    if ((Config().coreUrl != null) && (Config().appPlatformId != null) && (Config().coreOrgId != null) && (username != null) && (password != null)) {
+      String url = "${Config().coreUrl}/services/auth/login";
+      Map<String, String> headers = {
+        'Content-Type': 'application/json'
+      };
+      String? post = JsonUtils.encode({
+        'auth_type': auth2LoginTypeToString(usernameLoginType),
+        'app_type_identifier': Config().appPlatformId,
+        'api_key': Config().rokwireApiKey,
+        'org_id': Config().coreOrgId,
+        'creds': {
+          "username": username,
+          "password": password
+        },
+        'params': {
+          "sign_up": true,
+          "confirm_password": password
+        },
+        'profile': _anonymousProfile?.toJson(),
+        'preferences': _anonymousPrefs?.toJson(),
+        'device': deviceInfo,
+      });
+
+      Response? response = await Network().post(url, headers: headers, body: post);
+      if (response?.statusCode == 200) {
+        bool result = await processLoginResponse(JsonUtils.decodeMap(response?.body));
+        _notifyLogin(usernameLoginType, result);
+        return result ? Auth2UsernameSignUpResult.succeeded : Auth2UsernameSignUpResult.failed;
+      }
+      else if (Auth2Error.fromJson(JsonUtils.decodeMap(response?.body))?.status == 'already-exists') {
+        return Auth2UsernameSignUpResult.failedAccountExist;
+      }
+    }
+    return Auth2UsernameSignUpResult.failed;
+  }
+
+  Future<Auth2UsernameAccountState?> checkUsernameAccountState(String? username) async {
+    if ((Config().coreUrl != null) && (Config().appPlatformId != null) && (Config().coreOrgId != null) && (username != null)) {
+      String url = "${Config().coreUrl}/services/auth/account/exists";
+      Map<String, String> headers = {
+        'Content-Type': 'application/json'
+      };
+      String? post = JsonUtils.encode({
+        'auth_type': auth2LoginTypeToString(usernameLoginType),
+        'app_type_identifier': Config().appPlatformId,
+        'api_key': Config().rokwireApiKey,
+        'org_id': Config().coreOrgId,
+        'user_identifier': username,
+      });
+
+      Response? response = await Network().post(url, headers: headers, body: post);
+      if (response?.statusCode == 200) {
+        //TBD: handle Auth2EmailAccountState.unverified
+        return JsonUtils.boolValue(JsonUtils.decode(response?.body))! ? Auth2UsernameAccountState.exists : Auth2UsernameAccountState.nonExistent;
+      }
+    }
+    return null;
   }
 
   // Notify Login
@@ -1329,6 +1441,46 @@ Auth2OidcAuthenticateResult auth2OidcAuthenticateResultFromAuth2LinkResult(Auth2
     case Auth2LinkResult.succeeded: return Auth2OidcAuthenticateResult.succeeded;
     case Auth2LinkResult.failedAccountExist: return Auth2OidcAuthenticateResult.failedAccountExist;
     default: return Auth2OidcAuthenticateResult.failed;
+  }
+}
+
+// Auth2UsernameAccountState
+
+enum Auth2UsernameAccountState {
+  nonExistent,
+  exists,
+}
+
+// Auth2UsernameSignUpResult
+
+enum Auth2UsernameSignUpResult {
+  succeeded,
+  failed,
+  failedAccountExist,
+}
+
+Auth2UsernameSignUpResult auth2UsernameSignUpResultFromAuth2LinkResult(Auth2LinkResult value) {
+  switch (value) {
+    case Auth2LinkResult.succeeded: return Auth2UsernameSignUpResult.succeeded;
+    case Auth2LinkResult.failedAccountExist: return Auth2UsernameSignUpResult.failedAccountExist;
+    default: return Auth2UsernameSignUpResult.failed;
+  }
+}
+
+// Auth2UsernameSignInResult
+
+enum Auth2UsernameSignInResult {
+  succeeded,
+  failed,
+  failedNotFound,
+  failedInvalid,
+}
+
+Auth2UsernameSignInResult auth2UsernameSignInResultFromAuth2LinkResult(Auth2LinkResult value) {
+  switch (value) {
+    case Auth2LinkResult.succeeded: return Auth2UsernameSignInResult.succeeded;
+    case Auth2LinkResult.failedInvalid: return Auth2UsernameSignInResult.failedInvalid;
+    default: return Auth2UsernameSignInResult.failed;
   }
 }